requirements.txt
OpenEXR==1.3.2
Pillow==8.3.1
scipy==1.3.1
tensorboardX==1.8
torch==1.7.1
torchvision==0.8.2
tqdm==4.35.0
opencv-contrib-python==4.5.4.58
numpy==1.19.5
scikit-image==0.18.1